refactor: extract stake maturity modal
