+++ frontend/src/lib/components/neuron-detail/actions/NnsStakeMaturityButton.svelte
