+++ frontend/src/lib/components/sns-neuron-detail/actions/SnsStakeMaturityButton.svelte
@@ -1,13 +1,16 @@
 <script lang="ts">
   import { hasEnoughMaturityToStake } from "$lib/utils/sns-neuron.utils";
-  import {openSnsNeuronModal} from "$lib/utils/modals.utils";
+  import { openSnsNeuronModal } from "$lib/utils/modals.utils";
   import type { SnsNeuron } from "@dfinity/sns";
-  import {SELECTED_SNS_NEURON_CONTEXT_KEY, type SelectedSnsNeuronContext} from "$lib/types/sns-neuron-detail.context";
-  import {getContext} from "svelte";
+  import {
+    SELECTED_SNS_NEURON_CONTEXT_KEY,
+    type SelectedSnsNeuronContext,
+  } from "$lib/types/sns-neuron-detail.context";
+  import { getContext } from "svelte";
   import StakeMaturityButton from "$lib/components/neuron-detail/actions/StakeMaturityButton.svelte";
 
   const context: SelectedSnsNeuronContext =
-          getContext<SelectedSnsNeuronContext>(SELECTED_SNS_NEURON_CONTEXT_KEY);
+    getContext<SelectedSnsNeuronContext>(SELECTED_SNS_NEURON_CONTEXT_KEY);
   const { store }: SelectedSnsNeuronContext = context;
 
   let neuron: SnsNeuron | undefined | null;
@@ -16,8 +19,7 @@
   let enoughMaturity: boolean;
   $: enoughMaturity = hasEnoughMaturityToStake(neuron);
 
-  const showModal = () =>
-          openSnsNeuronModal({ type: "stake-maturity" });
+  const showModal = () => openSnsNeuronModal({ type: "stake-maturity" });
 </script>
 
 <StakeMaturityButton {enoughMaturity} on:click={showModal} />
